moved role change from util into its own submod
